Using rack-heartbeat gem to learn about Rack middleware.
The authors of that gem sort-of-but-didn't-really attempt to make it
usable as Rack middleware outside of a Rails app, but they use the
cattr_accessor method which is Rails-dependent, so this gem doesn't work
as-is. I had to use the facets gem which provides a collection add-on
functionality, including cattr_accessor.